Broke out tests for taking pieces into a separate class
EC, AH
