+++ src/test/java/mindchess/model/TestAlternatingTurns.java
@@ -33,40 +33,4 @@ public class TestAlternatingTurns {
         String currPlayer = model.getCurrentPlayerName();
         assertTrue(currPlayer.equals(p2) && !currPlayer.equals(p1));
     }
-
-    /**
-     * Tests that Pieces that are taken are saved in the deadPieces list
-     */
-    @Test
-    public void testTakePieceSuccess(){
-        IPiece IPiece = model.getCurrentBoardMap().get(new Square(3,7));
-
-        model.handleBoardInput(4,6);
-        model.handleBoardInput(4,5);
-
-        model.handleBoardInput(3,1);
-        model.handleBoardInput(3,2);
-
-        model.handleBoardInput(3,7);
-        model.handleBoardInput(6,4);
-
-        model.handleBoardInput(2,0);
-        model.handleBoardInput(6,4);
-
-        assertTrue(model.getCurrentDeadPieces().contains(IPiece));
-    }
-
-    /**
-     * Tests that if a piece is not taken, it does not get into the deadPieces list
-     */
-    @Test
-    public void testTakePieceFail(){
-        IPiece IPiece = model.getCurrentBoardMap().get(new Square(4, 0));
-
-        model.handleBoardInput(3,0);
-        model.handleBoardInput(4,0);
-
-        assertFalse(model.getCurrentDeadPieces().contains(IPiece));
-    }
-
 }
